OpenTelemetry/ExampleApplications/python/instrumented/pyproject.toml
[project]
name = "instrumented"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-sdk>=1.27.0",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.5",
]
